BLE: fix length problem...prevent CR in header!
The problem was not that the String was too long.  The problem was that the length of the string, once encoded into the header message resulted in one byte in the message being 0x0D, which is a CR.  Since the BLE module is expecting only ASCII characters from the message, it thinks that the CR is the end of the message, which is wrong, which breaks the system.  So, I sniff for a 0x0D and artificially add one byte to the message (a space character) to avoid this byte being 0x0D.