Removed forceToDataMode call.
This works as a hardware mechanism for switching to data mode, according to the Melody docs.
It doesn't seem to like RevE:V7 devices... Maybwe we have the wrong GPIO pin in our RevE definition?
PIO5 is listed as pin 2, but maybe it's 4?

Either way, the software way to swith modes for V5 and V6 and V7 is to send '$$$$' with a delay before and after. The BLE library should probably rely on this method anyway.